requirements.txt
mcp-use>=1.6.0
langchain-anthropic>=1.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
